+++ src/Tree/TaskNodeFactory.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Multitron\Tree;
+
+use Closure;
+
+interface TaskNodeFactory
+{
+    /**
+     * @param array<TaskNode> $children
+     * @param array<TaskNode|string> $dependencies
+     * @param string[] $tags
+     */
+    public function create(string $id, ?Closure $factory, array $children = [], array $dependencies = [], array $tags = [], ?string $class = null): TaskNode;
+}