+++ src/Tree/TaskTreeBuilderFactory.php
@@ -8,12 +8,12 @@ use Psr\Container\ContainerInterface;
 
 final readonly class TaskTreeBuilderFactory
 {
-    public function __construct(private ?ContainerInterface $container)
+    public function __construct(private ?ContainerInterface $container, private ?TaskNodeFactory $nodeFactory = null)
     {
     }
 
     public function create(): TaskTreeBuilder
     {
-        return new TaskTreeBuilder($this->container);
+        return new TaskTreeBuilder($this->container, $this->nodeFactory);
     }
 }